# Fix performance issue with deep nesting