@Davier 2021-02-03
While this PR fixes one issue with cache invalidation, there is still another issue causing exponential recursion.
I'll detail it here since I haven't been able to solve it yet.
The culprit seems to be the recursive call in the "Cross Size Determination". While experimenting on forcing the use of cached results, it reduced the number of recursion by 4 orders of magnitude (15'207'071 to 1'560 to be precise) in a test with about 20 nested containers. But doing that broke the cross size in fit-content mode (you can see it in this branch: https://github.com/Davier/stretch/commits/wip).
I'm sure there is a way to solve that issue, but it may need more than my basic understanding of the algorithm.